TIP-842: Move association type controller to structure sub context
