+++ src/Akeneo/Pim/Structure/Bundle/Controller/Ui/AssociationTypeController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Controller;
+namespace Akeneo\Pim\Structure\Bundle\Controller\Ui;
 
 use Akeneo\Pim\Structure\Component\Model\AssociationType;
 use Akeneo\Tool\Component\StorageUtils\Remover\RemoverInterface;
@@ -90,7 +90,7 @@ class AssociationTypeController
     public function editAction($code)
     {
         return [
-            'code'       => $code
+            'code' => $code
         ];
     }
 }
