+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/routing/external_api/association_type.yml
