+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/routing/front.yml
@@ -28,3 +28,9 @@ pim_enrich_family_edit:
     path: '/configuration/family/{code}/edit'
     requirements:
         code: '[a-zA-Z0-9_]+'
+
+pim_enrich_associationtype_edit:
+    path: '/configuration/association-type/{code}/edit'
+    defaults: { _controller: pim_enrich.controller.association_type:editAction }
+    requirements:
+        code: '[a-zA-Z0-9_]+'
