+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/routing/internal_api/associationtype.yml
@@ -1,12 +1,6 @@
 pim_enrich_associationtype_index:
     path: ''
 
-pim_enrich_associationtype_edit:
-    path: /{code}/edit
-    defaults: { _controller: pim_enrich.controller.association_type:editAction }
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
 pim_enrich_associationtype_rest_create:
     path: /rest
     defaults: { _controller: pim_enrich.controller.rest.association_type:createAction }
