+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/controllers.yml
@@ -7,7 +7,6 @@ parameters:
     pim_api.controller.media_file.class: Akeneo\Tool\Bundle\ApiBundle\Controller\MediaFileController
     pim_api.controller.currency.class: Akeneo\Tool\Bundle\ApiBundle\Controller\CurrencyController
     pim_api.controller.measure_family.class: Akeneo\Tool\Bundle\ApiBundle\Controller\MeasureFamilyController
-    pim_api.controller.association_type.class: Akeneo\Tool\Bundle\ApiBundle\Controller\AssociationTypeController
 
 services:
     pim_api.controller.category:
@@ -123,19 +122,4 @@ services:
             - '@pim_api.pagination.parameter_validator'
             - '@pim_api.pagination.offset_hal_paginator'
             - '%akeneo_measure.measures_config%'
-            - '%pim_api.configuration%'
-
-    pim_api.controller.association_type:
-        class: '%pim_api.controller.association_type.class%'
-        arguments:
-            - '@pim_api.repository.association_type'
-            - '@pim_serializer'
-            - '@pim_api.pagination.parameter_validator'
-            - '@pim_api.pagination.offset_hal_paginator'
-            - '@pim_catalog.factory.association_type'
-            - '@pim_catalog.updater.association_type'
-            - '@validator'
-            - '@router'
-            - '@pim_catalog.saver.association_type'
-            - '@pim_api.stream.association_type_partial_update_stream'
             - '%pim_api.configuration%'
