+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/routing.yml
@@ -18,10 +18,6 @@ currency:
     resource: '@PimApiBundle/Resources/config/routing/currency.yml'
     prefix: /rest/v1
 
-association_type:
-    resource: '@PimApiBundle/Resources/config/routing/association_type.yml'
-    prefix: /rest/v1
-
 fos_oauth_server_token:
     path: /oauth/v1/token
     defaults: { _controller: pim_api.controller.token:tokenAction }
