+++ src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -4,9 +4,7 @@ parameters:
     pim_enrich.controller.product.class:                      Pim\Bundle\EnrichBundle\Controller\ProductController
     pim_enrich.controller.product_model.class:                Pim\Bundle\EnrichBundle\Controller\ProductModelController
     pim_enrich.controller.group.class:                        Pim\Bundle\EnrichBundle\Controller\GroupController
-    pim_enrich.controller.association_type.class:             Pim\Bundle\EnrichBundle\Controller\AssociationTypeController
     pim_enrich.controller.file.class:                         Pim\Bundle\EnrichBundle\Controller\FileController
-    pim_enrich.controller.rest.association_type.class:        Pim\Bundle\EnrichBundle\Controller\Rest\AssociationTypeController
     pim_enrich.controller.rest.category.class:                Pim\Bundle\EnrichBundle\Controller\Rest\CategoryController
     pim_enrich.controller.rest.currency.class:                Pim\Bundle\EnrichBundle\Controller\Rest\CurrencyController
     pim_enrich.controller.rest.form_extension.class:          Pim\Bundle\EnrichBundle\Controller\Rest\FormExtensionController
@@ -86,18 +84,6 @@ services:
             - '@pim_enrich.form.group'
             - '@pim_catalog.factory.group'
 
-    pim_enrich.controller.association_type:
-        class: '%pim_enrich.controller.association_type.class%'
-        arguments:
-            - '@request_stack'
-            - '@router'
-            - '@translator'
-            - '@pim_catalog.repository.association'
-            - '@pim_enrich.form.handler.association_type'
-            - '@pim_enrich.form.association_type'
-            - '@pim_catalog.remover.association_type'
-            - '@pim_catalog.repository.association_type'
-
     pim_enrich.controller.file:
         class: '%pim_enrich.controller.file.class%'
         arguments:
@@ -120,18 +106,6 @@ services:
             - '@akeneo_batch_queue.manager.job_execution_manager'
             - '@oro_security.security_facade'
             - {'import': 'pim_importexport_import_profile_show', 'export': 'pim_importexport_export_profile_show'}
-
-    pim_enrich.controller.rest.association_type:
-        class: '%pim_enrich.controller.rest.association_type.class%'
-        arguments:
-            - '@pim_catalog.repository.association_type'
-            - '@pim_internal_api_serializer'
-            - '@pim_catalog.remover.association_type'
-            - '@pim_catalog.updater.association_type'
-            - '@pim_catalog.saver.association_type'
-            - '@validator'
-            - '@pim_user.context.user'
-            - '@pim_enrich.normalizer.violation'
 
     pim_enrich.controller.rest.group_type:
         class: '%pim_enrich.controller.rest.group_type.class%'
