+++ src/Pim/Bundle/EnrichBundle/Resources/config/routing.yml
@@ -42,10 +42,6 @@ group:
     resource: '@PimEnrichBundle/Resources/config/routing/group.yml'
     prefix: /enrich/group
 
-association:
-    resource: '@PimEnrichBundle/Resources/config/routing/associationtype.yml'
-    prefix: /configuration/association-type
-
 media:
     resource: '@PimEnrichBundle/Resources/config/routing/media.yml'
     prefix: /media
